backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
tensorflow==2.15.0
lightgbm==4.1.0
python-dotenv==1.0.0
requests==2.31.0
sqlite3